requirements.txt
flask
crewai
langchain
langchainhub
langchain-community
python-dotenv
groq
langchain-groq
